remove accessor getMoneySum :: Money a -> Sum a
We have an Iso for accessing the underlying value in a Money.  We
also removed the Iso for accessing the (Sum a) from a (Money a)
because the use case is not clear, and if it becomes clear it would
be better to provide a Control.Lens.Wrapped.Wrapped instance.

Therefore remove the getMoneySum accessor.